Refactor: FilterOption/DEFAULT_OPTION を共有モジュールに抽出し getAuthHeaders を共通化
レビュー指摘対応。
- FilterOption と DEFAULT_OPTION を client-safe な statsFilterOption.ts に抽出し、
  server(filterOptions.ts)と client(各コンテナ)で共有して重複定義のずれを防ぐ
- filterOptions.ts の getAuthHeaders を既存の services/v2/authHeaders に統一
- 認証/例外時のフォールバックを emptyOptions() 化し、共有参照の変異リスクを排除
